BULLET DODGED - the constrained recursive simulation in Simulator#evaluate was RETURNING its result, i.e. ALL the way up, including "goods"!
So once it starts a constrained search on one of the possibilities, it takes the first good one it finds, not evaluating the other possibilities. We want it to return (exit) on the first BAD it finds, but if it finds good, exit back to the previous level and continue evaluating choices.

Whew

modifies Simulator#evaluate to only return (ALL the way, remember!) if the result is bad